switch to ChunkCodecLibBlosc
