+++ filterpkgs/JLD2Blosc/Project.toml
@@ -4,9 +4,9 @@ authors = ["Jonas Isensee <jonas.isensee@web.de>"]
 version = "0.1.0"
 
 [deps]
-Blosc = "a74b3585-a348-5f62-a45c-50e91977d574"
+ChunkCodecLibBlosc = "c6a955be-ab7f-4fbb-b38f-caf93db6b928"
 JLD2 = "033835bb-8acc-5ee8-8aae-3f567f8a3819"
 
 [compat]
-Blosc = "0.7.3"
+ChunkCodecLibBlosc = "0.2.0"
 JLD2 = "0.5.15"
